refactor: separate filter operation from filter_entities mod
