+++ frontend/rust-lib/flowy-grid/src/services/filter/checkbox_filter.rs
@@ -1,5 +1,7 @@
 use crate::entities::{CheckboxCondition, GridCheckboxFilter};
-use crate::services::field::CheckboxCellData;
+use crate::services::cell::{AnyCellData, CellFilterOperation};
+use crate::services::field::{CheckboxCellData, CheckboxTypeOption};
+use flowy_error::FlowyResult;
 
 impl GridCheckboxFilter {
     pub fn apply(&self, cell_data: &CheckboxCellData) -> bool {
@@ -8,6 +10,16 @@ impl GridCheckboxFilter {
             CheckboxCondition::IsChecked => is_check,
             CheckboxCondition::IsUnChecked => !is_check,
         }
+    }
+}
+
+impl CellFilterOperation<GridCheckboxFilter> for CheckboxTypeOption {
+    fn apply_filter(&self, any_cell_data: AnyCellData, filter: &GridCheckboxFilter) -> FlowyResult<bool> {
+        if !any_cell_data.is_checkbox() {
+            return Ok(true);
+        }
+        let checkbox_cell_data: CheckboxCellData = any_cell_data.try_into()?;
+        Ok(filter.apply(&checkbox_cell_data))
     }
 }
 
@@ -22,6 +34,17 @@ mod tests {
             condition: CheckboxCondition::IsChecked,
         };
         for (value, r) in [("true", true), ("yes", true), ("false", false), ("no", false)] {
+            let data = CheckboxCellData(value.to_owned());
+            assert_eq!(checkbox_filter.apply(&data), r);
+        }
+    }
+
+    #[test]
+    fn checkbox_filter_is_uncheck_test() {
+        let checkbox_filter = GridCheckboxFilter {
+            condition: CheckboxCondition::IsUnChecked,
+        };
+        for (value, r) in [("false", true), ("no", true), ("true", false), ("yes", false)] {
             let data = CheckboxCellData(value.to_owned());
             assert_eq!(checkbox_filter.apply(&data), r);
         }
