+++ frontend/rust-lib/flowy-grid/src/services/filter/mod.rs
@@ -5,5 +5,6 @@ mod filter_service;
 mod number_filter;
 mod select_option_filter;
 mod text_filter;
+mod url_filter;
 
 pub(crate) use filter_service::*;
